Increase Coverage of Observable and ObservableList
