+++ app/src/main/java/com/github/sdpteam15/polyevents/database/observe/Observable.kt
@@ -1,11 +1,11 @@
 package com.github.sdpteam15.polyevents.database.observe
 
-import android.annotation.SuppressLint
-import androidx.arch.core.executor.ArchTaskExecutor
 import androidx.lifecycle.Lifecycle
 import androidx.lifecycle.LifecycleObserver
 import androidx.lifecycle.LifecycleOwner
 import androidx.lifecycle.OnLifecycleEvent
+import com.github.sdpteam15.polyevents.helper.HelperFunctions.observeOnStop
+import com.github.sdpteam15.polyevents.helper.HelperFunctions.run
 import java.util.*
 
 /**
@@ -45,18 +45,8 @@ class Observable<T>(value: T? = null) {
      *  @param observer lifecycle of the observer to automatically remove it from the observers when stopped
      *  @return a method to remove the observer
      */
-    fun observe(lifecycle: LifecycleOwner, observer: (T?) -> Unit): () -> Boolean {
-        val result = observe(observer)
-
-        //Anonymous class to observe the ON_STOP Event ao the Activity/Fragment
-        val lifecycleObserver = object : LifecycleObserver {
-            @OnLifecycleEvent(Lifecycle.Event.ON_STOP)
-            fun stopListener() = result()
-        }
-
-        lifecycle.lifecycle.addObserver(lifecycleObserver)
-        return result
-    }
+    fun observe(lifecycle: LifecycleOwner, observer: (T?) -> Unit): () -> Boolean
+        = observeOnStop(lifecycle, observe(observer))
 
     /**
      * Post a new value
@@ -68,14 +58,5 @@ class Observable<T>(value: T? = null) {
             for (obs in observers)
                 obs(value);
         })
-    }
-
-    @SuppressLint("RestrictedApi")
-    private fun run(runnable: Runnable) {
-        try {
-            ArchTaskExecutor.getInstance().postToMainThread(runnable)
-        } catch (e: RuntimeException) {
-            runnable.run()
-        }
     }
 }
