+++ app/src/main/java/com/github/sdpteam15/polyevents/helper/HelperFunctions.kt
@@ -1,9 +1,15 @@
 package com.github.sdpteam15.polyevents.helper
 
+import android.annotation.SuppressLint
+import androidx.arch.core.executor.ArchTaskExecutor
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.FragmentActivity
 import androidx.fragment.app.FragmentManager
 import androidx.fragment.app.FragmentTransaction
+import androidx.lifecycle.Lifecycle
+import androidx.lifecycle.LifecycleObserver
+import androidx.lifecycle.LifecycleOwner
+import androidx.lifecycle.OnLifecycleEvent
 import com.github.sdpteam15.polyevents.MainActivity
 import com.github.sdpteam15.polyevents.R
 
@@ -32,5 +38,24 @@ object HelperFunctions {
         val ft: FragmentTransaction = fragmentManager!!.beginTransaction()
         ft.setReorderingAllowed(false)
         ft.detach(frag).attach(frag).commit()
+    }
+
+    @SuppressLint("RestrictedApi")
+    fun run(runnable: Runnable) {
+        try {
+            ArchTaskExecutor.getInstance().postToMainThread(runnable)
+        } catch (e: RuntimeException) {
+            runnable.run()
+        }
+    }
+
+    fun observeOnStop(lifecycle: LifecycleOwner, result: () -> Boolean): () -> Boolean {
+        //Anonymous class to observe the ON_STOP Event ao the Activity/Fragment
+        val lifecycleObserver = object : LifecycleObserver {
+            @OnLifecycleEvent(Lifecycle.Event.ON_STOP)
+            fun stopListener() = result()
+        }
+        lifecycle.lifecycle.addObserver(lifecycleObserver)
+        return result
     }
 }
